Open URLs with a custom implementation in the platform module
Currently the underlying run script prints all its output, which can be
confusing if it produces output even on some successes. This will eventually
allow more control of how the output is presented, in addition to being closer
to how we'll need to implement file opening for custom file openers.