Parse custom commands in the feed config file
You can specify a command like

    command 'cmd here' unquoted @URL "double quoted" '@URL'

The command will apply to all feeds following the line, unless a different one
is set. You can remove a custom command and fall back to the default with just:

    command

on its own. The --open-with on the command line takes precedence over any
command specified in the config file.